actualizaciones y estructura para nueva fase del proyecto
