--- app/src/main/java/com/example/wherenow/ConfirmationScreen.kt
@@ -1,160 +1,0 @@
-package com.example.wherenow
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Favorite
-import androidx.compose.material3.*
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-import androidx.navigation.NavController
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
-import androidx.compose.ui.res.stringResource
-
-
-@Composable
-fun ConfirmationScreen(navController: NavController) {
-    val interests = listOf(
-        "Health & Wellness",
-        "Photography",
-        "Social & Networking"
-    )
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(
-                Brush.verticalGradient(
-                    colors = listOf(
-                        Color(0xFFB388FF),
-                        Color(0xFFFF80AB)
-                    )
-                )
-            ),
-        contentAlignment = Alignment.Center
-    ) {
-        Card(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(24.dp),
-            shape = RoundedCornerShape(24.dp),
-            elevation = CardDefaults.cardElevation(6.dp),
-            colors = CardDefaults.cardColors(containerColor = Color.White)
-        ) {
-            Column(
-                modifier = Modifier.padding(24.dp),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-
-                Text(
-                    text = stringResource(R.string.confirmation_title),
-                    style = MaterialTheme.typography.headlineSmall
-                )
-
-                Text(
-                    text = stringResource(R.string.confirmation_selected, interests.size),
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = Color.Gray
-                )
-
-                Spacer(Modifier.height(16.dp))
-
-
-                Icon(
-                    imageVector = Icons.Default.Favorite,
-                    contentDescription = "Heart",
-                    tint = Color(0xFF9C27B0),
-                    modifier = Modifier.size(48.dp)
-                )
-
-                Spacer(Modifier.height(16.dp))
-
-
-                Text(
-                    text = stringResource(R.string.confirmation_message),
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = Color.Gray
-                )
-
-                Spacer(Modifier.height(16.dp))
-
-
-                LazyVerticalGrid(
-                    columns = GridCells.Adaptive(minSize = 140.dp),
-                    horizontalArrangement = Arrangement.spacedBy(8.dp),
-                    verticalArrangement = Arrangement.spacedBy(8.dp),
-                    userScrollEnabled = false,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .heightIn(min = 0.dp, max = 200.dp)
-                ) {
-                    items(interests) { label ->
-                        InterestChip(label)
-                    }
-                }
-
-                Spacer(Modifier.height(24.dp))
-
-
-                Card(
-                    colors = CardDefaults.cardColors(containerColor = Color(0xFFF5F5F5)),
-                    shape = RoundedCornerShape(16.dp),
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    Column(modifier = Modifier.padding(16.dp)) {
-                    Text(
-                        stringResource(R.string.what_happens_next),
-                        style = MaterialTheme.typography.titleMedium
-                    )
-                    Spacer(Modifier.height(8.dp))
-                    Text("• " + stringResource(R.string.bullet_recommendations))
-                    Text("• " + stringResource(R.string.bullet_update))
-                    Text("• " + stringResource(R.string.bullet_discover))
-                    }
-                }
-
-                Spacer(Modifier.height(24.dp))
-
-
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    TextButton(onClick = { navController.popBackStack() }) {
-                        Text(stringResource(R.string.back))
-                    }
-                    Button(
-                        onClick = { navController.navigate(NavRoutes.AUTH) },
-                        colors = ButtonDefaults.buttonColors(containerColor = Color(0xFF9C27B0))
-                    ) {
-                        Text(stringResource(R.string.start_exploring))
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun InterestChip(label: String) {
-    Surface(
-        shape = RoundedCornerShape(50),
-        color = Color(0xFFEDE7F6),
-        modifier = Modifier.padding(horizontal = 4.dp)
-    ) {
-        Text(
-            text = label,
-            style = MaterialTheme.typography.bodySmall,
-            modifier = Modifier.padding(horizontal = 12.dp, vertical = 6.dp),
-            color = Color(0xFF512DA8)
-        )
-    }
-}
