+++ app/src/main/java/com/example/wherenow/NavRoutes.kt
@@ -4,5 +4,7 @@ object NavRoutes {
     const val AUTH = "auth"
     const val LOCATION = "location"
     const val QUIZ = "quiz"
-    const val CONFIRMATION = "confirmation"
+    const val HOME = "home"
+    const val SEARCH = "search"
+    const val EVENTS = "events"
 }
